refactor(fsi): remove unused FSI struct values
